Agrego lombok y organizo por carpetas
