+++ src/main/java/com/agenda/repositories/ContactoRepository.java
@@ -1,7 +1,7 @@
-package com.agenda;
+package com.agenda.repositories;
 
+import com.agenda.models.Contacto;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
